dashboard/requirements.txt
streamlit
pytest
pylint
pandas
psycopg2-binary
python-dotenv
altair
pyarrow
redshift_connector